refactor(parallel object): move to it's own new file
The classes of ParallelObject, ParallelObjectResult and ParallelObjectException,
are moved to it's own new file of sdcm/utils/parallel_object.py.
This will prevent some circular import issues.
