+++ performance_regression_operator_multi_tenant_test.py
@@ -14,7 +14,7 @@
 # Copyright (c) 2022 ScyllaDB
 
 from performance_regression_test import PerformanceRegressionTest
-from sdcm.utils.common import ParallelObject
+from sdcm.utils.parallel_object import ParallelObject
 from sdcm.utils.operator.multitenant_common import MultiTenantTestMixin
